% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almon.R
\name{pdl}
\alias{pdl}
\title{Generate polynomial disributed lags from a vector.}
\usage{
pdl(x, degree = 2, lags = 12)
}
\arguments{
\item{x}{a vector of time-ordered observations.}

\item{degree}{the degree of the polynomial. Default = 2.}

\item{lags}{the number of lags desired. Default = 12.}
}
\description{
Generate polynomial disributed lags from a vector.
}
\references{
Almon, Shirley. 1965. The distributed lag between capital 
       appropriations and net expenditures. Econometrica, 33, 178-196.
}
